requirement.txt
numpy<2.0.0
pycocoevalcap
huggingface_hub
transformers==4.46.3
accelerate==0.26.0
opencv-python
torch
Pillow

#RAG
faiss-cpu
faiss-gpu

#linting
autoflake
isort
flake8
black
pre-commit

opencv-python